+++ spec/bootstrap.php
@@ -1,0 +1,3 @@
+<?php
+// Nelmio\Alice uses a lot of finals which can get in the way of my specs
+DG\BypassFinals::enable();