+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Another fix for associating Alice v3 created entities with inline entities from EntityContext
 - Removed debug output used for investigating Alice issue which was fixed in 0.10.2.
+- When multiple fixture YAML files are loaded out of order, ensure all model associations are completed.
 
 ## [0.10.2] - 2020-10-01
 ### Fixed